+++ spec/braintree_spec.js
@@ -1,17 +1,30 @@
 describe("Braintree", function () {
+  var publicKey =
+    "MIIBCgKCAQEA8wQ3PXFYuBn9RBtOK3lW4V+7HNjik7FFd0qpPsCVd4KeiIfhuzupSevHUOLjbRSqwvAaZK3/icbBaM7CMAR5y0OjAR5lmmEEkc" +
+    "w+A7pmKQK6XQ8j3fveJCzC3MPiNiFfr+vER7O4diTxGhoXjFFJQpzKkCwFgwhKrW8uJLmWqVhQRVNphii1GpxI4fjFNc4h1w2W2CJ9kkv+9e3B" +
+    "nCpdVe1w7gBQZMkgjCzxbuAg8XaKlKD48M9kr8iE8kNt1eXV0jbmhCY3vZrckCUv26r2X4cD5lDvUtC1Gj6jBFobm/MelAfoFqNeq+/9VyMdYf" +
+    "hIecQimiBYr7Vm5VH9m69TXwIDAQAB";
+
   it("has a public encryption key", function () {
     var braintree = Braintree.create("a public key");
     expect(braintree.publicKey).toMatch("a public key");
   });
 
-  describe("encrypt", function () {
-    var publicKey =
-        "MIIBCgKCAQEA8wQ3PXFYuBn9RBtOK3lW4V+7HNjik7FFd0qpPsCVd4KeiIfhuzupSevHUOLjbRSqwvAaZK3/icbBaM7CMAR5y0OjAR5lmmEEkc" +
-        "w+A7pmKQK6XQ8j3fveJCzC3MPiNiFfr+vER7O4diTxGhoXjFFJQpzKkCwFgwhKrW8uJLmWqVhQRVNphii1GpxI4fjFNc4h1w2W2CJ9kkv+9e3B" +
-        "nCpdVe1w7gBQZMkgjCzxbuAg8XaKlKD48M9kr8iE8kNt1eXV0jbmhCY3vZrckCUv26r2X4cD5lDvUtC1Gj6jBFobm/MelAfoFqNeq+/9VyMdYf" +
-        "hIecQimiBYr7Vm5VH9m69TXwIDAQAB";
-    var badPublicKey = "MAUCAwEAAQ==";
+  it("encrypts with a known AES key", function() {
+    var braintree = Braintree.create(publicKey);
+    var key = Braintree.generateAesKey();
+    key.key = [-1958853821,973037212,398302960,1668515393,356607281,1582615236,558457690,-1305749256]
+    expect(key.encryptWithIv("test data", [1, 2, 3, 4])).toEqual("AAAAAQAAAAIAAAADAAAABJcSo857BMv+cJtJfpF5Pak=");
+  });
 
+  it("encrypts with another known AES key", function() {
+    var braintree = Braintree.create(publicKey);
+    var key = Braintree.generateAesKey();
+    key.key = [-1958853821,973037212,398302960,1668515393,356607281,1582615236,558457690,-1305749256]
+    expect(key.encryptWithIv("test data", [-320125304, 1018604835, -1358544460, 837200435])).toEqual("7OtGiDy2rSOvBkG0MeaqMzCOxy049Z8+7j33YHFKJ3E=");
+  });
+
+  describe("encrypt", function () {
     var publicExponent = "010001"
     var publicModulus = "00F304373D7158B819FD441B4E2B7956E15FBB1CD8E293B145774AA93EC09577829E8887E1BB3BA949EBC750E2E36D14AAC2F01A64ADFF89C6C168CEC2300479CB43A3011E659A610491CC3E03BA662902BA5D0F23DDFBDE242CC2DCC3E236215FAFEBC447B3B87624F11A1A178C5149429CCA902C0583084AAD6F2E24B996A958504553698628B51A9C48E1F8C535CE21D70D96D8227D924BFEF5EDC19C2A5D55ED70EE005064C9208C2CF16EE020F1768A94A0F8F0CF64AFC884F2436DD5E5D5D236E6842637BD9ADC90252FDBAAF65F8703E650EF52D0B51A3EA3045A1B9BF31E9407E816A35EABEFFD57231D61F84879C4229A2058AFB566E551FD9BAF535F"
     var privateExponent = "BCBFFD2C93CABC767684BBFF8ED52BCDED804AA5D1945CC6DE5BA9E1901459257122C97AA877448DB2282C711BA5F1CD7CA7DE0F32865F7B93190739EF8766880C321C1325D916F146E88D52ED95867426B6E2278C56B470B33081A8CBEE76FE9F5F6758AC7B909D2AD9A7FC7C33EBF691A0117C9EC51B4C11D915DE108EDEE4310343255FFEB52CAE06B2C93B8655DC370EA4E1F211B7DC40367377294AC6218B6668293D90AB17F07856F72AD59B45965C79F39BAA393BEDC178FFB77167BE3477912445445F020FA96B5D41AE9ED210ED2E90903CE4A5A6BE32247B2C4F6047A643DA1E2E8E9C62076B41C3EE26F27F07352908DCC00F2383839A6E64FB1"
@@ -58,11 +71,20 @@ describe("Braintree", function () {
       return Braintree.sjcl.codec.utf8String.fromBits(plaintext);
     }
 
-    it("handles malformed public keys", function () {
-      expect(function () {
-        var braintree = Braintree.create(badPublicKey)
-        braintree.encrypt("test data");
-      }).toThrow("Malformed public key");
+    describe("handles malformed public keys", function () {
+      it("when the exponent and modulus can't be extracted", function () {
+        expect(function () {
+          var braintree = Braintree.create("MAUCAwEAAQ==")
+          braintree.encrypt("test data");
+        }).toThrow("Invalid encryption key. Please use the key labeled 'Client-Side Encryption Key'");
+      });
+
+      it("when they can't be decoded", function () {
+        expect(function () {
+          var braintree = Braintree.create("36baf98243cc6709c3a60ba194881asd")
+          braintree.encrypt("test data");
+        }).toThrow("Invalid encryption key. Please use the key labeled 'Client-Side Encryption Key'");
+      });
     });
 
     it("encrypts the given text with the public key", function () {
@@ -83,21 +105,8 @@ describe("Braintree", function () {
     it("prepends the encrypted data with $bt4 and agent", function () {
       var braintree = Braintree.create(publicKey);
       var encrypted = braintree.encrypt("test data");
+
       expect(encrypted).toMatch(prefixRegExp);
-    });
-
-    it("encrypts with a known AES key", function() {
-      var braintree = Braintree.create(publicKey);
-      var key = braintree.generateAesKey();
-      key.key = [-1958853821,973037212,398302960,1668515393,356607281,1582615236,558457690,-1305749256]
-      expect(key.encryptWithIv("test data", [1, 2, 3, 4])).toEqual("AAAAAQAAAAIAAAADAAAABJcSo857BMv+cJtJfpF5Pak=");
-    });
-
-    it("encrypts with another known AES key", function() {
-      var braintree = Braintree.create(publicKey);
-      var key = braintree.generateAesKey();
-      key.key = [-1958853821,973037212,398302960,1668515393,356607281,1582615236,558457690,-1305749256]
-      expect(key.encryptWithIv("test data", [-320125304, 1018604835, -1358544460, 837200435])).toEqual("7OtGiDy2rSOvBkG0MeaqMzCOxy049Z8+7j33YHFKJ3E=");
     });
   });
 });